fix(tts): move prepare_tts_text to tts/__init__, use FrozenInstanceError in test
Move prepare_tts_text, _PAGINATION_RE, and _USER_PREFIX out of tts/base.py
into tts/__init__.py so base.py remains a pure Protocol/dataclass module
(matching whisper/base.py and llm/base.py). Update test import to
ccgram.tts and use FrozenInstanceError instead of bare Exception.
